Add landing page feature component
