+++ features/LandingPage/LandingPage.component.tsx
@@ -1,0 +1,9 @@
+import { HeroSection } from "./components";
+
+export default function LandingPageComponent() {
+  return (
+    <div className="min-h-screen relative">
+      <HeroSection />
+    </div>
+  );
+}
