+++ features/LandingPage/components/HeroSection/index.ts
@@ -1,0 +1,1 @@
+export { default as HeroSection } from "./HeroSection.component";
