+++ layout/AnimatedBackground/AnimatedBackground.layout.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { motion } from "motion/react";
 
 export default function AnimatedBackgroundLayout() {
